// games/simple-slots-shared/src/types/GameWsApiFieldsNames.ts
export enum GameWsApiFieldsNames {
    rId = '0',
    bid = '1',
    result = '2',
    status = '3',
    winAmount = '4',
    coefficient = '5',

    /** Кол-во вариантов на барабане */
    variantsCount = '6',
    /** Кол-во барабанов */
    reelsCount = '7',

    /** Сумма ставки */
    bidAmount = '8',
    balanceId = '9',
    currencyAddress = 'a',
    currencyNetwork = 'b',
}